+++ DATN_QLTiemChung/DATN_QLTiemChung_Api/Models/KhamSangLoc.cs
@@ -7,8 +7,8 @@ namespace DATN_QLTiemChung_Api.Models
         [Key]
         public string IDKS { get; set; }
         public string IDKH { get; set; }
-        public float ChieuCao { get; set; }
-        public float CanNang { get; set; }
+        public Double ChieuCao { get; set; }
+        public Double CanNang { get; set; }
         public DateTime ThoiGian { get; set; }
         public string IDNV { get; set; }
         public string TinhTrangSucKhoe { get; set; }
